[ALPS07847143] Disp: open aee layer,fix resume error
[Details]
open aee layer
revert bringup workround
fix cannot resume problem

MTK-Commit-Id: 52c5b5c24491957a67913506e8d54bc2810b1da1

CR-Id: ALPS07847143
Feature: display_mobile

+++ drivers/gpu/drm/mediatek/mediatek_v2/mtk_drm_drv.c
@@ -5360,7 +5360,7 @@ void mtk_drm_disp_test_init(struct drm_device *dev)
 		return;
 	}
 	DDPINFO("%s[%d] test_va=0x%llx, test_pa=0x%llx\n",
-		__func__, __LINE__, test_va, test_pa);
+		__func__, __LINE__, (unsigned long long)test_va, (unsigned long long)test_pa);
 
 	for (i = 0; i < height; i++)
 		for (j = 0; j < width; j++) {
